ratings/requirements.txt
asgiref==3.3.4
class-registry==2.1.2
Django==3.2.4
django-filter==2.4.0
djangorestframework==3.12.4
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
six==1.16.0
sqlparse==0.4.1
psycopg2==2.8
